# 🔧 Complete Lavalink Error Resolution: YAML + Apple Music Token Fixes