… testBot/APPLE_MUSIC_ERROR_SOLUTION.md
… testBot/COMPARISON_OLD_VS_NEW.md
… testBot/FINAL_GUARANTEE.md
… testBot/OAUTH_COOKIES_SETUP.md
… testBot/ULTIMATE_SUMMARY.md
… testBot/YAML_ERROR_SOLUTION.md
… testBot/YOUTUBE_PLAYBACK_GUARANTEE.md
… testBot/application-minimal.yml
… testBot/application-no-errors.yml
… testBot/application-optimized.yml
… testBot/application-ultimate-fixed.yml
… testBot/application-ultimate.yml
… testBot/application-working.yml
… testBot/cookies.txt
… testBot/setup-complete.sh
… testBot/start-lavalink-ultimate.sh
… testBot/youtube-playback-test.js